requirements.txt
asttokens==2.0.8
autopep8==2.0.1
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
jupyter_client==7.4.3
jupyter_core==4.11.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==21.3
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.3
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==304
pyzmq==24.0.1
six==1.16.0
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
tzdata==2023.3
wcwidth==0.2.5
